mks pointer math

+++ Marlin/src/lcd/extui/lib/mks_ui/draw_print_file.cpp
@@ -355,10 +355,10 @@ void disp_gcode_icon(uint8_t file_num) {
   #endif
 }
 
-uint32_t lv_open_gcode_file(char *path) {
+uintptr_t lv_open_gcode_file(char *path) {
   #if ENABLED(SDSUPPORT)
     uint32_t *ps4;
-    uint32_t pre_sread_cnt = 0;
+    uintptr_t pre_sread_cnt = 0;
     char *cur_name;
 
     cur_name = strrchr(path, '/');
@@ -368,7 +368,7 @@ uint32_t lv_open_gcode_file(char *path) {
     ps4 = (uint32_t *)strstr((char *)public_buf, ";simage:");
     // Ignore the beginning message of gcode file
     if (ps4) {
-      pre_sread_cnt = (uint32_t)ps4 - (uint32_t)((uint32_t *)(&public_buf[0]));
+      pre_sread_cnt = (uintptr_t)ps4 - (uintptr_t)((uint32_t *)(&public_buf[0]));
       card.setIndex(pre_sread_cnt);
     }
     return pre_sread_cnt;
