+++ Marlin/src/lcd/extui/lib/mks_ui/draw_ui.cpp
@@ -619,7 +619,7 @@ char *creat_title_text() {
     #if ENABLED(SDSUPPORT)
       //uint8_t re;
       //uint32_t read;
-      uint32_t pre_read_cnt = 0;
+      uintptr_t pre_read_cnt = 0;
       uint32_t *p1;
       char *cur_name;
 
@@ -630,7 +630,7 @@ char *creat_title_text() {
       p1 = (uint32_t *)strstr((char *)public_buf, ";simage:");
 
       if (p1) {
-        pre_read_cnt = (uint32_t)p1 - (uint32_t)((uint32_t *)(&public_buf[0]));
+        pre_read_cnt = (uintptr_t)p1 - (uintptr_t)((uint32_t *)(&public_buf[0]));
 
         To_pre_view              = pre_read_cnt;
         gcode_preview_over       = true;
@@ -666,7 +666,7 @@ char *creat_title_text() {
             uint32_t br  = card.read(public_buf, 400);
             uint32_t* p1 = (uint32_t *)strstr((char *)public_buf, ";gimage:");
             if (p1) {
-              gPicturePreviewStart += (uint32_t)p1 - (uint32_t)((uint32_t *)(&public_buf[0]));
+              gPicturePreviewStart += (uintptr_t)p1 - (uintptr_t)((uint32_t *)(&public_buf[0]));
               break;
             }
             else {
